requirements.txt
pandas
pyodbc
requests
python-dotenv
colorama
tabulate
flask
streamlit
sqlalchemy

